move states into corresponding module
