+++ packages/space-opera/src/components/camera_settings/components/pitch_limits.ts
@@ -18,8 +18,8 @@
 
 import {customElement, internalProperty} from 'lit-element';
 
-import {Camera} from '../../../redux/camera_state.js';
 import {registerStateMutator, State} from '../../../redux/space_opera_base.js';
+import {Camera} from '../camera_state.js';
 import {Limits} from '../types.js';
 
 import {LimitsBase} from './limits_base.js';
