+++ packages/space-opera/src/components/ibl_selector/initial_environment_images.ts
