+++ packages/space-opera/src/components/ibl_selector/lighting_state.ts
