+++ packages/space-opera/src/components/materials_panel/material_state.ts
