+++ packages/space-opera/src/redux/space_opera_base.ts
@@ -19,12 +19,11 @@ import {GltfModel, ModelViewerConfig} from '@google/model-viewer-editing-adapter
 import {ModelViewerElement} from '@google/model-viewer/lib/model-viewer';
 import * as Redux from 'redux';  // from //third_party/javascript/redux:redux_closurized
 
+import {Camera, INITIAL_CAMERA} from '../components/camera_settings/camera_state.js';
 import {HotspotConfig} from '../components/hotspot_panel/hotspot_config.js';
+import {INITIAL_ENVIRONMENT_IMAGES} from '../components/ibl_selector/initial_environment_images.js';
+import {EnvironmentImage} from '../components/ibl_selector/lighting_state.js';
 import {getGltfEdits, GltfEdits, INITIAL_GLTF_EDITS} from '../components/model_viewer_preview/gltf_edits.js';
-
-import {Camera, INITIAL_CAMERA} from './camera_state.js';
-import {INITIAL_ENVIRONMENT_IMAGES} from './initial_environment_images.js';
-import {EnvironmentImage} from './lighting_state.js';
 
 /**
  * Space Opera state.
