+++ packages/space-opera/src/redux/space_opera_base_test.ts
@@ -20,10 +20,10 @@ import {arrayBufferEqualityTester, createBufferFromString, generatePngBlob, GlTf
 import {cloneJson} from '@google/model-viewer-editing-adapter/lib/util/clone_json.js'
 import {RGBA} from '@google/model-viewer/lib/model-viewer';
 
+import {applyCameraEdits, Camera} from '../components/camera_settings/camera_state.js';
 import {dispatchAddBaseColorTexture, dispatchAddEmissiveTexture, dispatchAddMetallicRoughnessTexture, dispatchAddNormalTexture, dispatchAddOcclusionTexture, dispatchBaseColorTexture, dispatchEmissiveTexture, dispatchMaterialBaseColor, dispatchMetallicFactor, dispatchNormalTexture, dispatchOcclusionTexture, dispatchRoughnessFactor, dispatchSetAlphaCutoff, dispatchSetAlphaMode, dispatchSetEmissiveFactor} from '../components/materials_panel/reducer.js';
 import {applyEdits, generateTextureId, getGltfEdits, INITIAL_GLTF_EDITS} from '../components/model_viewer_preview/gltf_edits.js';
 
-import {applyCameraEdits, Camera} from './camera_state.js';
 import {dispatchCurrentCameraState, dispatchGltfAndEdits, dispatchGltfUrl, dispatchInitialCameraState, reduxStore, registerStateMutator} from './space_opera_base.js';
 
 const EXAMPLE_BIN_AS_STRING = 'example of some bin data';
